Fix [https://trac.osgeo.org/mapguide/ticket/575]. Add x64 configurations to some key foundation components.
git-svn-id: http://svn.osgeo.org/mapguide/trunk@3185 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Foundation/System/MemoryStreamHelper.cpp
@@ -218,7 +218,7 @@ MgStreamHelper::MgStreamStatus MgMemoryStreamHelper::GetSingle(float& data, bool
     float fData = 0.0;
 
     MgStreamHelper::MgStreamStatus stat = GetData(&fData, sizeof(fData), blocking, peeking);
-    data = MG_NTOHL(fData);
+    data = (float)MG_NTOHL(fData);
 
     return stat;
 };
