+++ MgDev/Oem/ACE/ACE_wrappers/ace/Log_Msg.h
@@ -649,7 +649,7 @@ private:
   static u_long flags_;
 
   /// Offset of msg_[].
-  static int msg_off_;
+  static ptrdiff_t msg_off_;
 
   /**
    * Number of existing ACE_Log_Msg instances; when 0, delete program/host
