+++ MgDev/Oem/DWFTK7.1/develop/global/src/dwf/expat/xmlparse.c
@@ -1581,7 +1581,7 @@ long XMLCALL
 XML_GetCurrentByteIndex(XML_Parser parser)
 {
   if (eventPtr)
-    return parseEndByteIndex - (parseEndPtr - eventPtr);
+    return parseEndByteIndex - (long)(parseEndPtr - eventPtr);
   return -1;
 }
