+++ MgDev/Oem/dbxml-2.3.10/xqilla/src/context/impl/XQDynamicContextImpl.cpp
@@ -229,12 +229,12 @@ void XQDynamicContextImpl::setExternalContextNode(const XERCES_CPP_NAMESPACE_QUA
   setContextItem(new NodeImpl(node, this));
 }
 
-void XQDynamicContextImpl::setContextSize(unsigned int size)
+void XQDynamicContextImpl::setContextSize(size_t size)
 {
   _contextSize = size;
 }
 
-void XQDynamicContextImpl::setContextPosition(unsigned int pos)
+void XQDynamicContextImpl::setContextPosition(size_t pos)
 {
   _contextPosition = pos;
 }
@@ -244,12 +244,12 @@ Item::Ptr XQDynamicContextImpl::getContextItem() const
   return _contextItem;
 }
 
-unsigned int XQDynamicContextImpl::getContextSize() const
+size_t XQDynamicContextImpl::getContextSize() const
 {
   return _contextSize;
 }
 
-unsigned int XQDynamicContextImpl::getContextPosition() const
+size_t XQDynamicContextImpl::getContextPosition() const
 {
   return _contextPosition;
 }
